+++ RectAdapter.cpp
@@ -8,22 +8,22 @@ RealRect RectAdapter::frectToReal(const FilterRect& frect,
 	if (frect.getHBeg() == 0)
 		real.hBeg = 0;
 	else
-		real.hBeg = image.h / frect.getHBeg();
+		real.hBeg = (unsigned int)(image.h / frect.getHBeg());
 
 	if (frect.getWBeg() == 0)
 		real.wBeg = 0;
 	else
-		real.wBeg = image.w / frect.getWBeg();
+		real.wBeg = (unsigned int)(image.w / frect.getWBeg());
 
 	if (frect.getHEnd() == 0)
 		real.hEnd = 0;
 	else
-		real.hEnd = image.h / frect.getHEnd();
+		real.hEnd = (unsigned int)(image.h / frect.getHEnd());
 
 	if (frect.getWEnd() == 0)
 		real.wEnd = 0;
 	else 
-		real.wEnd = image.w / frect.getWEnd();
+		real.wEnd = (unsigned int)(image.w / frect.getWEnd());
 
 	return real;
 }
